revise assignTimestamps and assignTimestampsModal to add modularity for question types
