+++ web/src/components/modals/questionTypes/MultipleChoice.vue
@@ -1,0 +1,90 @@
+<template>
+    <div>
+        <div v-if="activities.length >= 1 && activities[activityIndex] != ''" class="activity-info">
+            <label for="question-text">Question Text: </label><input type="text" id="question-text" name="question-text" :value="activities[activityIndex].questionText">
+            <label for="question-text">Answer 1: </label><input type="text" id="answer-one" name="answer-one" :value="activities[activityIndex].answers[0]" :onchange="populateSelectOptions">
+            <label for="question-text">Answer 2: </label><input type="text" id="answer-two" name="answer-two" :value="activities[activityIndex].answers[1]" :onchange="populateSelectOptions">
+            <label for="answers">Correct Answer: </label>
+            <select id="answers" name="answers">
+                <option v-for="(answer,index) in activities[activityIndex].answers" :key="answer" v-bind:value="answer" :selected="answer === activities[activityIndex].correctAnswer" v-bind:id="optionIds[index]">{{answer}}</option>
+            </select>
+        </div>
+        <div v-else class="activity-info">
+            <label for="question-text">Question Text: </label><input type="text" id="question-text" name="question-text">
+            <label for="question-text">Answer 1: </label><input type="text" id="answer-one" name="answer-one" :onchange="populateSelectOptions">
+            <label for="question-text">Answer 2: </label><input type="text" id="answer-two" name="answer-two" :onchange="populateSelectOptions">
+            <label for="answers">Correct Answer: </label>
+            <select id="answers" name="answers">
+                <option id="option-one" value=""></option>
+                <option id="option-two" value=""></option>
+                {{activityIndex}}
+            </select> 
+        </div>
+        <div class="button-div">
+            <div class="save"><button type="button" class="btn-green" @click="save()">Save</button></div>
+            <div class="close"><button type="button" class="btn-green" @click="close()">Close</button></div>
+        </div>
+        <p id="invalid-save"></p>
+    </div>
+</template>
+
+<script>
+export default {
+    name: 'MultipleChoice',
+    data() {
+        return {
+            activityModalArray: [],
+            allInputsValid: false,
+            activitySaved: false,
+            optionIds: ['option-one','option-two']
+        }
+    },
+    props: {
+        activities: Array,
+        activityIndex: Number
+    },
+    methods: {
+        close() {
+            this.$emit('close')
+        },
+        save() {
+            let invalid = document.getElementById('invalid-save')
+            this.checkInputs()
+            if(this.allInputsValid) {
+                invalid.innerHTML = ""
+                this.activitySaved = true
+                this.$emit('save',this.activityModalArray)
+            } else {
+                invalid.innerHTML = "Please fill out all fields."
+                this.activityModalArray = []
+            }
+        },
+        checkInputs() {
+            let questionText = document.getElementById('question-text')
+            let questionOne = document.getElementById('answer-one')
+            let questionTwo = document.getElementById('answer-two')
+            let option = document.getElementById('answers')
+            if(questionText.value != "" && questionOne.value != "" && questionTwo.value != "") {
+                this.allInputsValid = true
+                this.activityModalArray = [questionText.value,questionOne.value,questionTwo.value,option.value]
+            } else {
+                this.allInputsValid = false
+            }
+        },
+        populateSelectOptions() {
+            let optionOne = document.getElementById('option-one')
+            let optionTwo = document.getElementById('option-two')
+            let answerOne = document.getElementById('answer-one').value
+            let answerTwo = document.getElementById('answer-two').value
+            optionOne.value = answerOne
+            optionOne.innerHTML = answerOne
+            optionTwo.value = answerTwo
+            optionTwo.innerHTML = answerTwo
+        }
+    }
+}
+</script>
+
+<style scoped>
+
+</style>
